refactor: ExchangeRequestPage에서 ExchangePostDetail 컴포넌트 분리
